http: widen the per-request id to 64 bits
The id that routes abort, request-body write, receive-resume and
cert-check-resume messages from the JS thread to a request's socket on
the HTTP thread was a u32 counter. Make it u64 everywhere it is stored
or compared (counter, AsyncHTTP/HTTPClient, abort tracker key, HTTP
thread messages, h2 by_http_id index and pending-connect scans, h3
session scans, S3 download task) and start the counter at 1 so 0 is
only ever the value of requests without a signal store.

Adds fetchRequestIdInternals.skipIds to bun:internal-for-testing so a
test can hold two live requests exactly 2^32 ids apart and check that
they are aborted independently.

+++ src/http/HTTPContext.rs
@@ -408,7 +408,7 @@ impl<const SSL: bool> HTTPContext<SSL> {
     /// misses: a request parked in `PendingConnect.waiters` (coalesced
     /// onto a leader's in-flight TLS connect) never registered a socket,
     /// so it can only be found by scanning here.
-    pub(crate) fn abort_pending_h2_waiter(&mut self, async_http_id: u32) -> bool {
+    pub(crate) fn abort_pending_h2_waiter(&mut self, async_http_id: u64) -> bool {
         if !SSL {
             return false;
         }
