+++ src/http/h3_client/ClientContext.rs
@@ -195,7 +195,7 @@ impl ClientContext {
         session.registry_index = u32::MAX;
     }
 
-    pub(crate) fn abort_by_http_id(async_http_id: u32) -> bool {
+    pub(crate) fn abort_by_http_id(async_http_id: u64) -> bool {
         let Some(this) = Self::get() else {
             return false;
         };
@@ -213,7 +213,7 @@ impl ClientContext {
         false
     }
 
-    pub(crate) fn stream_body_by_http_id(async_http_id: u32, ended: bool) {
+    pub(crate) fn stream_body_by_http_id(async_http_id: u64, ended: bool) {
         let Some(this) = Self::get() else {
             return;
         };
@@ -227,7 +227,7 @@ impl ClientContext {
         }
     }
 
-    pub(crate) fn resume_receive_by_http_id(async_http_id: u32) {
+    pub(crate) fn resume_receive_by_http_id(async_http_id: u64) {
         let Some(this) = Self::get() else {
             return;
         };
